google-sheet-web-scraper/requirements.txt
# Google Sheet Web Scraper Requirements

# Google Sheets API
gspread>=6.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Web Scraping
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# WebDriver (optional - can also download manually)
webdriver-manager>=4.0.0
